# docs/swagger.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: "/api"
info:
  version: "2.0.0"
  title: Jomon API
  description: >-
    JomonのAPIです。
...
paths:
...
  /transactions:
    get:
      operationId: "getTransactions"
      description: トランザクション一覧を取得する。
      tags:
        - Transactions
      parameters:
        - $ref: "#/components/parameters/sortTransactionQuery"
        - $ref: "#/components/parameters/targetQuery"
        - $ref: "#/components/parameters/sinceQuery"
        - $ref: "#/components/parameters/untilQuery"
        - $ref: "#/components/parameters/limitQuery"
        - $ref: "#/components/parameters/offsetQuery"
        - $ref: "#/components/parameters/tagQuery"
        - $ref: "#/components/parameters/groupQuery"
        - $ref: "#/components/parameters/requestQuery"
      responses:
        "200":
          description: 該当するものがない場合は空配列を返却。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
    post:
      operationId: postTransaction
      description: トランザクションを新規作成する。管理者権限が必要。
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostTransaction"
      responses:
        "201":
          description: 作成に成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
  /transactions/{transactionID}:
    parameters:
      - name: transactionID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getTransactionDetail
      description: 指定したトランザクションの詳細を取得する。
      tags:
        - Transactions
      responses:
        "200":
          description: あったら返す。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/404"
    put:
      operationId: putTransactionDetail
      description: 指定したTransactionを修正する。管理者権限が必要。requestに紐づいている場合は変更不可。
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostTransactionWithOneTarget"
      responses:
        "200":
          description: OK 修正後のトランザクションを返却する。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
...
components:
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "2020講習会"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - created_at
        - updated_at
...
    Group:
      type: object
      nullable: true
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "SysAd"
        description:
          type: string
          example: "SysAd班"
        budget:
          type: integer
          nullable: true
          example: 250000
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - budget
        - created_at
        - updated_at
...
    Target:
      type: string
      example: "hoge株式会社"
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
          example: 1200
        target:
          $ref: "#/components/schemas/Target"
        request:
          type: string
          nullable: true
          format: uuid
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        group:
          nullable: true
          $ref: "#/components/schemas/Group"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - amount
        - target
        - request
        - tags
        - group
        - created_at
        - updated_at
    PostTransaction:
      type: object
      properties:
        amount:
          type: integer
          example: 1200
        targets:
          type: array
          items:
            $ref: "#/components/schemas/Target"
        tags:
          type: array
          items:
            type: string
            format: uuid
        group:
          type: string
          nullable: true
          format: uuid
        request:
          type: string
          nullable: true
          format: uuid
      required:
        - amount
        - targets
        - request
        - tags
        - group
    PostTransactionWithOneTarget:
      type: object
      properties:
        amount:
          type: integer
          example: 1200
        target:
          $ref: "#/components/schemas/Target"
        tags:
          type: array
          items:
            type: string
            format: uuid
        group:
          type: string
          nullable: true
          format: uuid
        request:
          type: string
          nullable: true
          format: uuid
      required:
        - amount
        - targets
        - request
        - tags
        - group
...
  responses:
...
    400:
      description: 不正なリクエスト。
    403:
      description: 編集権限がない人による操作。
    404:
      description: 指定したリソースは存在しない。
...
  parameters:
...
    sortTransactionQuery:
      name: sort
      description: 並び順 (作成日時が新しい "created_at", 作成日時が古い "-created_at", 金額の昇順 "amount", 金額の降順 "-amount")
      required: false
      in: query
      schema:
        type: string
    targetQuery:
      name: target
      description: 誰との取引か
      required: false
      in: query
      schema:
        type: string
        format: uuid
    sinceQuery:
      name: since
      description: いつからの依頼か
      required: false
      in: query
      schema:
        type: string
        format: date
    untilQuery:
      name: until
      description: いつまでの依頼か
      required: false
      in: query
      schema:
        type: string
        format: date
    limitQuery:
      name: limit
      description: 取得する最大個数
      required: false
      in: query
      schema:
        type: integer
        minimum: 0
        default: 100
    offsetQuery:
      name: offset
      description: スキップする個数
      required: false
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    tagQuery:
      name: tag
      description: タグ(複数の場合カンマ区切り)
      required: false
      in: query
      schema:
        type: string
    groupQuery:
      name: group
      description: グループ
      required: false
      in: query
      schema:
        type: string
    requestQuery:
      name: request
      description: 依頼
      required: False
      in: query
      schema:
        type: string
        format: uuid